// frontend/src/api/generated/models/mergeDecisionOutcome.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type MergeDecisionOutcome = typeof MergeDecisionOutcome[keyof typeof MergeDecisionOutcome];


export const MergeDecisionOutcome = {
  Merged: 'Merged',
  KeptSeparate: 'KeptSeparate',
  Reverted: 'Reverted',
} as const;